Re-use subscriber when re-fetching
The typical scenario when subscribing and calling FETCH is that
they either get a message or will block. And if the block, they
will timeout after a few seconds. What they typically do in that
case is to re-fetch on the same subscription.

We don't really know when a client times out, so we can't remove
them from the list of blocked subscribers when that happens. But
we can at least re-use their entry in the list of blocked clients
when they call FETCH again.